requirements.txt
Flask==3.0.3
flask_cors==5.0.0
PyJWT==2.9.0
PyMySQL==1.1.1
pytest==8.3.3
python-dotenv==1.0.1
python_bcrypt==0.3.2